+++ priv/testrunner/vendor.js
@@ -1,0 +1,5 @@
+import ErlangTypes from 'erlang-types';
+
+export default {
+  ErlangTypes,
+};