+++ src/javascript/lib/core.js
@@ -8,6 +8,7 @@ import lists from './core/erlang_compat/lists';
 import elixir_errors from './core/erlang_compat/elixir_errors';
 import elixir_config from './core/erlang_compat/elixir_config';
 import io from './core/erlang_compat/io';
+import filename from './core/erlang_compat/filename';
 import binary from './core/erlang_compat/binary';
 import unicode from './core/erlang_compat/unicode';
 import Store from './core/store';
@@ -28,6 +29,18 @@ function get_global() {
 
   console.warn('No global state found');
   return null;
+}
+
+function initApp() {
+  const Elixir = {};
+
+  Elixir.__table__ = {};
+  Elixir.start = (app, args) => {
+    app.__load(Elixir).start(Symbol.for('normal'), args);
+  };
+  Elixir.load = module => module.__load(Elixir);
+
+  return Elixir;
 }
 
 const globalState = get_global();
@@ -52,9 +65,11 @@ export default {
   lists,
   elixir_errors,
   io,
+  filename,
   binary,
   unicode,
   elixir_config,
   math,
   proplists,
+  initApp,
 };