+++ test/cli_test.exs
@@ -10,7 +10,7 @@ defmodule ElixirScript.CLI.Test do
   test "process help" do
     assert capture_io(fn ->
       ElixirScript.CLI.process(:help)
-    end) =~ "usage: elixirscript <module> [options]"
+    end) =~ "usage: elixirscript <module | path> [options]"
   end
 
   test "process version" do
@@ -22,12 +22,12 @@ defmodule ElixirScript.CLI.Test do
   test "process unknown" do
     assert capture_io(fn ->
       ElixirScript.CLI.process({"", [unknown: ""]})
-    end) =~ "usage: elixirscript <module> [options]"
+    end) =~ "usage: elixirscript <module | path> [options]"
   end
 
   test "process input" do
     assert capture_io(fn ->
       ElixirScript.CLI.process({["Atom"], []})
-    end) =~ "export default Elixir"
+    end) =~ "export default"
   end
 end